rename: `Home` 컴포넌트 파일 위치 수정
